+++ app/src/index.ts
@@ -1,16 +1,11 @@
-import { app, BrowserWindow, ipcMain } from 'electron';
+import { app, BrowserWindow } from 'electron';
 const axios = require('axios');
 declare const MAIN_WINDOW_WEBPACK_ENTRY: any;
 
 // Keep a reference for dev mode
-let devMode = false;
-
-// Determine the mode (dev or production)
-if (process.defaultApp ||
-   /[\\/]electron-prebuilt[\\/]/.test(process.execPath) || 
-   /[\\/]electron[\\/]/.test(process.execPath)) {
-  devMode = true;
-}
+let devMode = (process.defaultApp ||
+  /[\\/]electron-prebuilt[\\/]/.test(process.execPath) || 
+  /[\\/]electron[\\/]/.test(process.execPath));
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) { // eslint-disable-line global-require
@@ -60,24 +55,3 @@ app.on('activate', () => {
 
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and import them here.
-if (devMode) {
-  const dotenv = require('dotenv');
-  dotenv.config();
-}
-
-// pass settings path to renderer process
-ipcMain.on('settings-path', (event, arg) => {
-  event.returnValue = process.env.AV_SETTINGS_PATH;
-})
-
-// pass initial data paths to renderer process
-ipcMain.on('data-paths', (event, arg) => {
-  let dataPaths: string[] = [];
-  Object.keys(process.env).forEach((key: string) => {
-    if (key.startsWith('AV_DATA_PATH')) {
-      dataPaths.push(process.env[key]);
-    }
-  });
-
-  event.returnValue = dataPaths;
-})